CA-768: refactor(app-header): extract AppHeader feature module; make ShellModule navigation-only
