+++ test/app/shell/widgets/app_shell_page_app_bar_test.dart
@@ -1,7 +1,7 @@
 import 'package:construculator/app/shell/app_shell_bloc/app_shell_bloc.dart';
 import 'package:construculator/app/shell/app_shell_page.dart';
 import 'package:construculator/app/shell/shell_module.dart';
-import 'package:construculator/app/shell/widgets/header_row.dart';
+import 'package:construculator/features/app_header/presentation/widgets/header_row.dart';
 import 'package:construculator/features/dashboard/dashboard_module.dart';
 import 'package:construculator/features/dashboard/presentation/bloc/dashboard_bloc/dashboard_bloc.dart';
 import 'package:construculator/features/dashboard/presentation/bloc/project_dropdown_bloc/project_dropdown_bloc.dart';
@@ -96,7 +96,6 @@ void main() {
           ),
         ],
         child: AppShellPage(
-          projectUIProvider: Modular.get<ProjectUIProvider>(),
           currentProjectNotifier: Modular.get<CurrentProjectNotifier>(),
           router: Modular.get<AppRouter>(),
         ),
