+++ test/features/app_header/screenshots/goldens/header_row/390.0x64.0/header_row_with_username.png
