+++ test/features/app_header/screenshots/goldens/notification_icon/56.0x56.0/notification_icon_no_badge.png
