+++ test/features/app_header/screenshots/goldens/profile_avatar/56.0x56.0/profile_avatar_letter_j.png
