+++ test/features/app_header/screenshots/header_row_screenshot_test.dart
@@ -1,4 +1,4 @@
-import 'package:construculator/app/shell/widgets/profile_avatar.dart';
+import 'package:construculator/features/app_header/presentation/widgets/header_row.dart';
 import 'package:construculator/l10n/generated/app_localizations.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -7,17 +7,17 @@ import '../../../utils/screenshot/await_images_extension.dart';
 import '../../../utils/screenshot/font_loader.dart';
 
 void main() {
-  const size = Size(56, 56);
+  const size = Size(390, 64);
   const ratio = 1.0;
-  const testName = 'profile_avatar';
+  const testName = 'header_row';
   TestWidgetsFlutterBinding.ensureInitialized();
 
   setUpAll(() async {
     await loadAppFontsAll();
   });
 
-  group('ProfileAvatar Screenshot Tests', () {
-    testWidgets('renders letter avatar for name starting with J', (
+  group('HeaderRow Screenshot Tests', () {
+    testWidgets('renders default state with no notifications and no user', (
       tester,
     ) async {
       tester.view.physicalSize = size;
@@ -29,8 +29,9 @@ void main() {
           locale: const Locale('en'),
           localizationsDelegates: AppLocalizations.localizationsDelegates,
           supportedLocales: AppLocalizations.supportedLocales,
-          home: const Scaffold(
-            body: Center(child: ProfileAvatar(name: 'John')),
+          home: Scaffold(
+            appBar: HeaderRow(),
+            body: const SizedBox.shrink(),
           ),
         ),
       );
@@ -38,16 +39,14 @@ void main() {
       await tester.awaitImages();
 
       await expectLater(
-        find.byType(ProfileAvatar),
+        find.byType(HeaderRow),
         matchesGoldenFile(
-          'goldens/$testName/${size.width}x${size.height}/${testName}_letter_j.png',
+          'goldens/$testName/${size.width}x${size.height}/${testName}_default.png',
         ),
       );
     });
 
-    testWidgets('renders letter avatar for name starting with A', (
-      tester,
-    ) async {
+    testWidgets('renders with unread notification badge', (tester) async {
       tester.view.physicalSize = size;
       tester.view.devicePixelRatio = ratio;
 
@@ -57,8 +56,9 @@ void main() {
           locale: const Locale('en'),
           localizationsDelegates: AppLocalizations.localizationsDelegates,
           supportedLocales: AppLocalizations.supportedLocales,
-          home: const Scaffold(
-            body: Center(child: ProfileAvatar(name: 'Alice')),
+          home: Scaffold(
+            appBar: HeaderRow(unreadNotificationCount: 5),
+            body: const SizedBox.shrink(),
           ),
         ),
       );
@@ -66,14 +66,14 @@ void main() {
       await tester.awaitImages();
 
       await expectLater(
-        find.byType(ProfileAvatar),
+        find.byType(HeaderRow),
         matchesGoldenFile(
-          'goldens/$testName/${size.width}x${size.height}/${testName}_letter_a.png',
+          'goldens/$testName/${size.width}x${size.height}/${testName}_with_badge.png',
         ),
       );
     });
 
-    testWidgets('renders letter avatar when imageUrl is empty', (
+    testWidgets('renders with username for profile letter avatar', (
       tester,
     ) async {
       tester.view.physicalSize = size;
@@ -85,8 +85,9 @@ void main() {
           locale: const Locale('en'),
           localizationsDelegates: AppLocalizations.localizationsDelegates,
           supportedLocales: AppLocalizations.supportedLocales,
-          home: const Scaffold(
-            body: Center(child: ProfileAvatar(name: 'Bob', imageUrl: '')),
+          home: Scaffold(
+            appBar: HeaderRow(userName: 'John Doe'),
+            body: const SizedBox.shrink(),
           ),
         ),
       );
@@ -94,9 +95,9 @@ void main() {
       await tester.awaitImages();
 
       await expectLater(
-        find.byType(ProfileAvatar),
+        find.byType(HeaderRow),
         matchesGoldenFile(
-          'goldens/$testName/${size.width}x${size.height}/${testName}_empty_url.png',
+          'goldens/$testName/${size.width}x${size.height}/${testName}_with_username.png',
         ),
       );
     });
