+++ test/features/app_header/widgets/accessibility/title_search_app_bar_a11y_test.dart
@@ -1,4 +1,4 @@
-import 'package:construculator/app/shell/widgets/profile_avatar.dart';
+import 'package:construculator/features/app_header/presentation/widgets/title_search_app_bar.dart';
 import 'package:construculator/l10n/generated/app_localizations.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -11,29 +11,33 @@ void main() {
     await loadAppFontsAll();
   });
 
-  Widget makeTestableWidget({ThemeData? theme, String name = 'John'}) {
+  Widget makeTestableWidget({ThemeData? theme}) {
     return MaterialApp(
       theme: theme ?? createTestTheme(),
       locale: const Locale('en'),
       localizationsDelegates: AppLocalizations.localizationsDelegates,
       supportedLocales: AppLocalizations.supportedLocales,
       home: Scaffold(
-        body: Center(child: ProfileAvatar(name: name)),
+        appBar: TitleSearchAppBar(onSearchTap: () {}),
+        body: const SizedBox.shrink(),
       ),
     );
   }
 
-  group('ProfileAvatar – accessibility', () {
+  group('TitleSearchAppBar – accessibility', () {
     testWidgets(
-      'meets label guideline for letter avatar in both themes',
+      'search button meets tap target and label guidelines in both themes',
       (tester) async {
         await setupA11yTest(tester);
-        // Tap target size check is disabled: Figma specifies 40×40 for the
-        // profile avatar, which is below the 48dp Android / 44pt iOS minimum.
+        // Tap target size is disabled: the container's vertical padding
+        // squeezes the inner AppBar to 40px, so the search icon measures
+        // 48x40 — a pre-existing layout carried over verbatim from
+        // AppShellPage. TODO: [CA-822] rework the layout to reach 48x48.
+        // https://ripplearc.youtrack.cloud/issue/CA-822
         await expectMeetsTapTargetAndLabelGuidelinesForEachTheme(
           tester,
           (theme) => makeTestableWidget(theme: theme),
-          find.byKey(const Key('profile_avatar_button')),
+          find.byKey(const Key('title_search_app_bar_search_button')),
           checkTapTargetSize: false,
         );
       },
