+++ test/features/app_header/widgets/header_row_test.dart
@@ -1,6 +1,6 @@
-import 'package:construculator/app/shell/widgets/header_row.dart';
-import 'package:construculator/app/shell/widgets/notification_icon.dart';
-import 'package:construculator/app/shell/widgets/profile_avatar.dart';
+import 'package:construculator/features/app_header/presentation/widgets/header_row.dart';
+import 'package:construculator/features/app_header/presentation/widgets/notification_icon.dart';
+import 'package:construculator/features/app_header/presentation/widgets/profile_avatar.dart';
 import 'package:construculator/l10n/generated/app_localizations.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
